Start attaching the chip identification info.
In order to actually do register IO I need to get the driver to the point
where it attaches the register offsets for things.  That's coming next.

Now, this stuff is badly organised - I'm going to just get it up to the
point of running and then tidy things up a bit before I make a start on
BMI/HIF/CE support.