Begin fleshing out the copy engine, pipe and descdma code.
This doesn't yet fully compile!

I'm making somewhat of a choice to BSD-ify it rather than keeping
it mostly linux.  I don't mind for now; the code is in small enough
pieces that converting it /back/ won't be that hard.

Start copying the descdma code from ath(4); the eventual plan is to
somewhat abstract the busdma memory abuse out so (a) it's debuggable
and (b) it should be slightly easier to port things.