Re-enable interrupts once we handle them.
Turns out they were being disabled as part of the wait for target process,
and .. this isn't useful for what we're doing.

This results in the normal legacy IRQ path working fine. I'm seeing more
interrupts that I'm not handling via the CE path for some reason and I'll
have to figure that out, but ... hello QCA9880v2.  It's really nice to
see you again.