Start migration net80211 code out of being called during attach and do it where ath10k does - after the core is probed.
In order to have the MAC address available we need to do the net80211
setup /after/ we've successfully woken up the firmware and received
a completion notification. (Copying it in will come next so we can
use it when initialising net80211.)

But, a callout is still registered and firing, sigh.  Hopefully it
is the RX HTT pipe refill thing I had commented out; so enable that
and retest.

Add a stub drain_tx too, as we will need to do it soon too.