Migrate the athp_vap code over to ath10k_vif.
The rest of the driver uses that.

Next up is athp_node; I don't know what I'm going to do there yet.
It's a hold-over from the otus/rsu driver base I used to get the
net80211 pieces up and I'll have to replace it with an ath10k
specific struct.