Finish making the interface creation/deletion code compile.
It hasn't yet been tested at all, and the monitor setup path had to be
stubbed out.  Ideally monitor mode should be the first mode to test
as it will stress the receive path out and should (!) be the easiest
device to implement, but .. well, we'll see.  First this has to not
panic when it's called.