Track if the ath10k linux code did indeed set things up correctly or not.
Don't try to tear down the ath10k_vif state if we didn't successfully
set it up.  Otherwise it gets highly confused.

This is a completely dirty hack to just try to cope with the interface
creation path not currently working.  It fails one of the WMI commands;
need to dig into it a bit more.