the ath10k upstream driver actually reinits locks and callouts, tasks, etc each time ath10k_core_start() is called.
Kalle thinks this is a bug, but it's not fatal on linux .. but it is on
freebsd.

So this is pass #1 on making sure that doesn't happen.
It's currently enough to ensure ath10k_start() doesn't kernel panic
on doing bad things to locks, etc - but it still fails on the
second pass of HTC init.  It's getting an unexpected message (0x3).
I'll figure that out next.