Intermediary state of the minimal requirements to try management frame xmit.
Firstly - need to bring up enough to call the htt_mgmt_tx routine.
For native wifi, we need to juggle the QoS bits out of the frame
as appropriate to appease the microsoft gods.  Since we're using
native wifi right now and 802.3, we have to the juggling here.

(If/when it comes time to use 802.3 - we still have to use 802.11
frames for the "raw" path, so we'll still have to convert 802.11
to whatever the firmware encap path is.  Then for normal data path
it'll be 802.3, and we hopefully won't need to modify them.
Anyway, baby steps.)

For freeing tx mbufs, hopefully (!) we can use the exact same mbuf
given to us by net80211, complete with TX callback information.
That way the existing machinery in net80211 will hopefully just
work.  Don't implement that bit just yet - add a printf() so when
I come to this part I'll quickly implement it.

This doesn't yet compile - there is a lot of missing stuff to fill in.
I'll do that next.