Start implementing tx flushing.
Which is "wait until we've finish transmitting before we do more work."

We have to make sure that the transmit queue is flushed before we
delete peers.  Otherwise if we get a data frame queued out without
an active peer, TX hangs.

Now, mac80211 is more in control of when transmit queues start/stop.
net80211 currently leaves it up to the drivers.  So, we're likely
going to have to add some more of that logic to this ath10k driver
to make sure that peers exist when frames are queued, and ensuring
when peers are deleted we temporarily hold traffic and wait for
HTT to flush.

Ugh!  This needs better debugging!