Begin some more crypto, station, etc stuff.. Yes - god, I should've done this as separate commits.
* change the channel stuff to frequency uint32_t's, and '0' means "none"
* start setting the current channel when we get the foreign channel
  notification - since that's how we track off-channel traffic like
  scan results
* start adding crypto encap stuff - untested yet; since trying a wpa
  assocation causes the RX path to see encap'ed assoc resp, etc
  frames and I'm not sure why.
* don't set CAPINFO_PRIVACY if we're not yet associated? Not sure
  if this is "right" yet.
* make dwell time in scan configurable and set it to 200ms for now -
  the 50ms passive scans aren't good enough right now so we don't
  pick up everything.
* add a new flag from net80211 (uncommitted) - saying we're an offload
  scan device, so don't do things like probe requests, powersave in net80211, etc
* flesh out some of the scan offload methods, like iwm does.

The net80211 channel list "list chan" isn't the same channel set that
the firmware is programmed in with.  I'm not sure yet; maybe we're getting
that channel list before we get our regulatory domain set.  So, some of the
scan results/frames come back with a NULL channel lookup as net80211
doesn't know about DFS channels.  Will dig further.

Still trying to see why/whether we're seeing and missing scan results or not.