This is a big commit that addresses a variety of dma tag related issues and so far has quietened busdma/iommu hilarity.
* add some map/unmap debugging
* make most things 4-byte aligned, not 8-byte.
  the dmar code gets unhappy and will complain if you pass in a dword
  aligned mbuf into an 8 byte aligned tag - it doesn't copy it for you
  via bounce buffers or anything, it just plainly complains.
* .. leave the things we need 8 byte aligned (notably RX buffers, CE
  slots) alone
* add a lock around map/unmap - this seems to quieten the iommu code
  from complaining about things.  more later.
* add a missing unload - saw this after a code review with ath10k
* comment out the per pci copyengine ring dmatag - we weren't using it.

Now, this stops the IOMMU from asserting on out of bounds operations.
What seemed to happen here is:

* map mbuf A, get paddr X
* map a second mbuf B, get paddr X too for some reason
* unmap mbuf B, paddr X
* then the DMAR throws an error
* then unmap mbuf A, paddr X - at the correct point.

The busdma code doesn't have any locks in the bounce buffer or dmar
code, so my guess is that it expects the driver code to provide
locking around map/unmap as appropriate.  I'm not yet sure about
whether the lock is required around busdma memory allocation and
free - that'll come next.  I think as long as everything is NOWAIT,
I should be able to wrap it with a lock and everything will work
out fine.

TODO:

* wrap the busdma alloc/free calls with the DMA lock as well