First pass at fixing up this driver locking, unfinished.
The target is to try and remove the sleep lock and instead convert
it to a normal mutex and fix all the places where we hold a lock
when we shouldn't.

The initial pass is to convert the conf lock to a mutex instead of
a sleep lock.  The sleep lock makes a bunch of locking problems
kinda go away, but it makes other things more complicated.
Notably, since we need to sleep for certain activities in interrupt
mode (ie in interrupt handlers), we, well, can't.

This may not be the final solution, but it does indeed allow me to
attempt to tidy up how locking is handled.

The no-nos include:

* the tsleep() in the compl wait paths, which so far haven't complained
  with other mutexes
* holding the lock across taskqueue_cancel / taskqueue_drain
* holding the lock across some busdma operations (tag_create, memalloc)
  where we shouldn't

Note: this may still prove to be problematic.  We'll see.