This is a whole lot of hacks, most we will need in some fashion but it's .. not stopping NIC hangs.
mac80211 seems to stop the NIC after mac80211 has stopped.
FreeBSD doesn't do it, so we call ath10k_stop when the last
vap is torn down.  However, there are pending calls that net80211
makes which reaches back into the driver and those need to stop.

This mostly stops the panicing when the driver is unloaded, but i'm
sure a whole lot more odd stuff is there to fix.

However, this doesn't really do much yet - a bit of parallel TCP will
hang QCA9880v2, and I am totally not yet sure why.