Add in stubbed out interrupt handling code from ath10k/pci.c.
This should eventually take place of the interrupt code that
I dreamt up when first bringing the driver up.

Note: the linux interrupt stuff starts as a filter and it then
schedules a tasklet like other tasklets.  Maybe that processing
is why we don't get as much lock contention / preemption in Linux.
In FreeBSD the same mechanism would begin as a filter and schedule
an ithread, which can be preempted by other work including
the taskqueue we are scheduling work on.