ath10k_stop() frees the HTT TX locks and such.  We need it to happen /after/ the VAP is destroyed, otherwise the NIC is down when we are trying to do things like tear down peers / flush TX queues once the locks are freed.
Now, maybe the "right" thing to do here is to have net80211 trigger
off a driver task that is effectively "we're done, you can shut down
now" and have /that/ be what we hang ath10k_stop() off of.
That way it can be serialised with the rest of the net80211 state
changes (ioctls, packets, etc.)

This fixes a crash when one does 'ifconfig wlan1 destroy' with
an active wpa_supplicant session.  Still, there's more to do..