Experiment by deferring the BSS update notification path.
This currently breaks STA mode operation.

The STA mode BSS updates need to be synchronous as part of the
call to newstate, including making sure the BSS peer is there.

It may be that for STA mode we need to do updates as part of the
newstate call, but for AP mode we need to do updates deferred.

Those experiments will come next!