Partially (mostly) undo the STA side of the peer setup code for station operation.
Why?  Because I'm mostly lazy.

Well, it really is because of a reasonably simple problem - if too
much work is deferred, then certain things .. well, should also be deferred.
For example, if the STA BSS update is deferred, then the association
frame is sent before the STA BSS frame is sent, and things don't work.

So, ideally there should be a way of temporarily delaying transmit
frames until the state update has happened.  In particular, it would be good
for the following:

* pause transmit until a keycache program has completed - this is
  needed for the final EAPOL frame exchange which requires being
  encrypted by the newly programmed key.

  (Note: the upstream firmware does this for you, so if you do encryption
  but you don't setup an open key, offload transmit stalls.  lol.)

  (Note: the CT firmware has stripped out the above firmware code,
  so if it works it's because we won the race.)

* pause transmit until a node has been programmed in and the BSS
  programming is completed.  This is done for us here because we're
  doing STA updates sync in newstate, but .. yeah, this is broken
  for AP mode.

This at least fixes STA mode enough to start using the NIC again.
It doesn't fix the hostap buffering stuff; that'll have to be fixed.