Add some debugging to try and understand why beeliner attach isn't working.
It turns out it's not working because beeliner wants to attach more MSI
interrupts (12 CE, versus 7 for peregrine/rome) and this confuses our
allocation routine.

I'll look into it in some more depth.  It's possible that this mechanism
just plain used to fail in linux ath10k from last year.