Initial changes to replumb keys after the STA BSS comes active.
STA mode operation requires the BSS keys to be replumbed after the BSS
comes active.  Net80211 will call key_set when the key is set, which
typically is before the association occurs.  So, we do indeed track
setting the WEP keys up, but we fail to program them until the association
is done.

This also sets the default key index to something.

It's not perfect - it doesn't work for >1 key just yet (again due to
default tx key index.)  that'll come next.