Initial hack to break out the RX completion from RX work itself.
This does remove the lock contention and doing the buffer free
from inside the HTT_RX lock, but it doesn't (yet) improve performance.

All of the lock contention is gone though! The only obvious lock
contention now is in the TCP stack.

So, likely the rest of the issues are to be found in TX queue and
RX queue handling.. more to come.