Temporary hack for enabling hostap operation for testing.
For now, to allow for hostap operation, we need to bring up
the NIC early because the initial group key is programmed in
for some reason BEFORE the first transition from INIT -> RUN.
I think that's a net80211 stupidity, but I'll address that
later.

Without it, the first GTK would not be programmed in, so hostapd
operation fails.