Updates to sync to what linux ath10k does!  at least for station mode!
The rough path is:

* vdev create
* (stuff)
* vdev scan

then to auth:

* vdev start
* create peer
* send frames

then to assoc:

* vdev stop, vdev chanchange, vdev start if reqired
* vdev associate with associate info

Change the flow to do this.  In particular, we need a peer, but we
don't need a temporary association when sending the auth frame, which
means we then don't need to do a vif restart.

I need to go double/triple-check the channel upgrade path during
association (and for AP operation) to see if I detect a channel/width/opmode
change and then cycle through a restart.

TODO:

* So far so good.  I'm missing a bunch of stuff, notably QoS, beacon interval,
  powersave config, slot time, etc.  That'll come next.

* I'm also doing a vdev down after a scan where I don't technically need to be.
  I'll check that next.