Start improving the debugging around the vdev start/restart path and if it's being done wrong.
I've noticed that beacon interval and dtim is set to 0 in the first
association pass here; and not reset when the association is torn down /
new one is done.  This is likely an ordering issue (ie, the arvif cached
parameters are not being initialised before vdev start/restart) and ..
yeah, I need to fix that.