This is a WIP for experimenting around the driver transmit/receive completion path.
ath10k uses tasklets to complete things like transmit and receive which is cheaper
on linux because they're guaranteed to be serialised after the caller and on the
same CPU.

However on FreeBSD the taskqueue can run on any CPU, including running immediately,
which means although our latency may be less for starting the deferred work,
it is occuring /whilst/ we're actively trying to finish what we're doing.
This results in a lot of interesting lock contention - the biggest being the
RX work (that's already deferred from the RX HTT notification) which then itself
defers refilling the HTT RX rings.

This set of work attempts to not waste time deferring things and grabbing/releasing
locks in inefficient spots.  I'm going to do a bit more tweaking to see how much
better I can get this before I ask for any further help.