New athp arvif lock (#9)
* changed timeout to 2 seconds

for sloweer systems I have changed the ATH10K_HTC_WAIT_TIMEOUT_MSEC to 2 seconds instead.

* took back the change.

Instead i'll implement a system that unloads the firmware when the capabilities don't come in correctly and the firmware load crashes. kldload and unload should not have to deal with internal driver issues.

* add code to force the driver to attempt to retry to load the interface up to 6 times.

add code to force the driver to attempt to retry to load the interface up to 6 times. This should fix kldload issues, as long as the wifi card will work during boot we can automate bringing up the interface in scripts. Thats our first goal.

* re-order

* re-trying probe fw until sucessful or errors at the 6th try.

* undo changes

* attempt at retrying to wait for htc to complete and return

stopping hif and restarting hif if htc wait target doesnt finish succesfully.

* removed return on hif_stop

* attemping to recall ath10k_core_probe_fw

trying to reset the atheros chip to to load up properly on error.

* add sleep statement

sleep a little to let the device catch up and be ready.

* dumb auto programming features of VSCode capatilized Sleep

* not use to freebsd kernel developing.

I decided to use pause_sig in the case a signal called us to release before we end sleeping.

* added in a ath10k_core_stop

need to stop the core its not being called on error. I also have set bmi.done_sent = false.

* oops forgot semicolon

* test removing ath10k_core_stop from retrying probe and just hack bmi.done_sent

* cleaened up my hack for probe retrying

turned the probe retry in to functional code and added a retry limit to core.h called ATH10K_FW_PROBE_RETRYS

* forgot struct def

* added function to cleanup memory leaks

added a function to the process that allows us to cleanup any memory leaks if there are any.

* removed pause_sig

pause_sig isn't being used anywhere else and pause_sig is a delay based function which does not put the thread in the sleep queue so I switched it to tsleep instead.

* removed line of code from ath10k

* add null check.

Grammer correction and add a null check for fwlog locking mutext variable as well as adding a longer timeout for htc connect.

* used wrong case

used wrong case for null, woops so use to the many other programming languages with lowercase null.

* added some mutex checking

added mutext checking inside of if_athp_fwlog for some reason its still trying to log on an unitialized mutex, which means somewhere something is still initialized for the first firmware failed load, need to check this out in detail, its weird.

* forgot &

* chaned from 3 seconds to 1 didnt do anything.

* removing annoying warnings

removing annoying warnings that come out of my serial device, so i cant see or type anything.

* Revert "removing annoying warnings"

This reverts commit ad31d5de1230980d37ffff771ca8d1ec9086d0ad.

* Revert "Revert "removing annoying warnings""

This reverts commit b535d1482bf2541c45f487c0e7e7a60d3e30755d.

* first commit

Initialize commit on trying to move the hostap buffer allocation to the athp_parent instead of the add interface section.

* fixed type error

wrong type was used.

* remove extra pointers

remove extra pointers.

* out of order

moved functions in the right order.

* forgot initializer

* formatting

* removed call to deallocate inside of allocate and the label.

removed call to deallocate inside of allocate and the label.

* added static for compile via git on phone

* added dereference via git on phone

* moved dma allocate to attach.

* moved dma allocate in to if_athp_pci attach

* required to add mac.h

* moved beacon allocate before the start of all the task queues.

* changed from pointer to object.

* pointer change didn't make it.

* remove assignment address from non_pointer

* fixed issue with issue #3 and the htt_rx_ring_refill

ath10k_htt_rx_ring_refill does a dma allocate so no other lock can be held during this processor that can't be slept, I found that wmi_tx_beacons was holding the conf lock so I created a wait for that.

* Create README.md

* change default power to 8 dbM

tx_power_recalc takes txpower and * 2 checking to see if lowering increases thw power level of tx because it seems as if tx is not being set because its value is too high.

* changed power level to 9 which equals 18.

18 is the power level for below:
11ac HT40	MCS0	18	-93

* moved ath10k_wmi_wait_for_tx_beacons_ready

moved ath10k_wmi_wait_for_tx_beacons_ready to
ath10k_wmi_cmd_send

* ATHP_ARVIF_LOCK

changed all conf lock that involved the arvif array to a new arvif lock instead.

* changed to conf lock

* changed arvif lock for txpower

changed the locks on txpower to arvifs lock instead of conf lock.

* removed a few not needed arvif locks and changed a few locks.

* Revert "removed a few not needed arvif locks and changed a few locks."

This reverts commit 2a664da0e89892d677dc4498f3c94717756d51e1.

* Revert "changed arvif lock for txpower"

This reverts commit c24e4c2eb30991c15a97481aa03c6c26a6158252.

* fixed locks/kernel crash via git on phone

* Update README.md

* Update README.md

* Update README.md

* Update README.md