Initial UAPSD/WME updates.
I think this is correct for both hostap and station mode operation.
It unblocks setting up uapsd if it's configured in hostap or station
modes and I'm doing some further tinkering with it before moving
onto the next thing (likely crypto key deletion with a now-freed
ieee80211_key reference. ugh.)