[athp] Fix the SKB RX hash function for in-order receive buffer tracking.
This path is only used by the later QCA6174 hardware (hw 3.2), not the
earlier QCA6174 hardware.

This fixes a boot-time panic when the driver loads, but there's still
some issues sending WMI commands to this particular hardware revision.

My guess is that things aren't fully initialised enough for HTC to
operate after the firmware is loaded and the BMI response occurs.