[athp] Fix the phymode configuration when associating nodes in AP mode.
I'm testing an AR9380 STA (HT20) on a 5GHz VHT20 channel AP.
Because the phymode was set to ic_curchan, this is VHT20
and the firmware thus tries VHT rates that just fail.
The per-node phymode needs to be the node channel information,
NOT the AP channel.

This should also be fine in STA mode as the BSS channel should
be set A-OK.

(cherry picked from commit f825f230bc7c89fe293fd9936480395266f54dd0)