[athp] Fix firmware startup + VAP running tracking; remove current restart logic
* remove the current restart logic; I think we should reimplement
  it in a slightly less recursive way and make it go through the
  ATH10K_STATE logic.

* Don't use sc_isrunning for whether firmware is running or not -
  this breaks when creating the first interface (which starts firmware,
  adds interfaces via WMI but doesn't start them) and doing the first
  ifconfig up.

Before this change, the first VAP would load firmware and set sc_isrunning=1;
but when athp_parent() is called it wouldn't call ieee80211_start_all()
to kick all the VAP state machines.

Now ar->state is used to see if the firwmare is up or not, which is what
it's for.  sc_isrunning is now exclusively used for tracking whether
we've started VAPs or not.