athp: cleanup modules build Makefiles
The Makefiles had a lot of copy and paste and especially CLEANFILES
was wrong in most of them.
Rather than keeping the duplicate logic everywhere, put it into
a common Makefile.inc and only have the per-firmware kmod specifics
in the individual Makefiles.
This will also allow us to change the way we are building the
firmware(9) kernel modules in a single place.

Sponsored by:	Rubicon Communications, LLC (d/b/a "Netgate")