Start methodising the is_in_peer_table stuff so I can doubly check locking.
So, quick background.

net80211 doesn't do any buffering/queueing in the network stack between
adding a node and transmitting a frame.  It expects that adding a node,
allocating/setting/changing a key, etc are all immediate things that
don't sleep - and then once the call returns, a frame can be transmitted.

This is true for like, fullmac devices like ath(4), but it's definitely
not true for firmware things.  So, athp and a lot of other drivers just
push those updates into a deferred queue.  Trouble is, they may then
just transmit stuff with a not-yet-programmed peer, and some hardware
may or may not get angery at you for doing.

So yes, although net80211 needs to grow this support to make firmware
NICs a LOT easier to implement, for now we're stuck having to track
this ourselves.  Hence, is_in_peer_table and its various checks.

The checks aren't enforced or methodized yet - I have to go convert
them to use the CONF lock everywhere.  But setting it between 1 and
0 was easy to do, so it's done here.