[athp] untested commit against -HEAD, along with a local net80211 change
This is just a compile test on a dev machine, I haven't yet tested
it with an actual QCA9880 NIC.  I'll do that soon.

There's a change to add ic_vht_conf for VHT40/VHT80; I've landed it here
to tinker with it, but I haven't yet landed THAT diff in net80211.
So just stay tuned, comment out those two lines if needed.