[athp] add a TODO about TKIP key length and format
This expects the key + MIC to be formatted the same as net80211,
which expects the net80211 key layout to not change when key sizes
grow.

Put TODOs here, I need to clean up net80211's key stuff to expose
the key and MIC via accessors, and NOT making assumptions about
the key layout!