[athp] clarify what I need to do for regulatory and DFS domain stuff.
The DFS domain stuff is a bit sticky - we currently don't enumerate the
specific DFS domains in regdomain.xml; I need to figure out how to
add that and expose it via net80211 so I don't have to hard-code the
behaviour here.

The regulatory stuff requires a little more careful reading of the
ath10k/ath_regd code in Linux to understand what the WMI parameters
are and the EEPROM codes, and if they're effectively the same thing.