[athp] Placeholder regulatory domain routines
Implement a pair of regulatory domain routines for getting the EEPROM
value / calculating the initial setup and then fetching the current
config.

There's a lot missing here, but the goal of this work is to move
the regulatory stuff into regd/ routines, and have if_athp_mac.c
just call it.  That way I can remove the TODOs there and then
focus on the rest of the regulatory stuff.