[mtwn] continue poking at firmware loading
I'm getting closer to enumerating all of the routines required
to load in firmware, and it's quite a trip.

The fun/unfun part here is where to put the code that shovels
in the firmware.  Putting USB transfer code in mt7610/usb/ seems a
bit..  problematic.

The actual firmware load bits, which I haven't yet done, also do
blocking bulk tranfers on one of the bulk endpoints to shovel in
the firmware.  I don't know if I can get away with that here?

Anyway, I'm still chipping away at it..