[mtwn] migrate the USB MCU stuff into usb/mt7610u (and the if_mtwn_usb module)
I was not liking the direction the MCU bring-up code was going, because
it was starting to need to do a whole bunch of USB vendor and bulk endpoint
work.

So instead now I'm going to migrate it into usb/ , where the rest of the
USB bulk endpoint / vendor control stuff lives.  It should be enough
to be able to do the USB shenanigans required.