[mtwn] implement an MVP MCU reponse parser
* flip on waiting for all commands, so Q_SELECT also populates
  a non-zero seq field
* implement enough of the RX path to process an RX buffer and attempt
  to parse it.

This demonstrates it working end to end.  Next is figuring out how
to create a command "link" between sent commands and their responses,
so I can get the responses for stuff like the MCU memory reads.

(Right now my "wait" is just waiting for the transfer to complete;
it's NOT waiting on the command completion itself!)