[mtwn] implement waiting for RX notification for completion
* refactor out how we sleep on mtwn_cmd buffers so it can be called
  from a few places
* add methods to add and complete waiting mtwn_cmd buffers, so
  the MCU RX completion method can parse the response and
  complete the right thing.
* And .. well, do the above.
* Split the "wait" field in msg_send into two - one that waits for
  buffer TX, and one that waits for the RX completion to process.
* Set the cmd->seq field when assigning a seq number to request
  an RX completion (yes I'm terrible for not using a method ...)

I modified the mcu function select routine to request a completion
notification from firmware to test this path out, as the path I need
it for (RX MCU register reads, both single and regpair lists) aren't
yet implemented.  It seems to work fine.

The reason the flag is split is so that things that don't generate
RX notifications - such as loading firmware - don't try waiting
for a notification that won't succeed.