[mtwn] begin adding the BBP initvals and switch table setup
This is still a WIP, but progress is happening.

init_bbp() is programming in three tables - the middle one is
the switch table which needs some conditional programming.
Next is that switch table programming.

Some calibration info is required for this, so add in some calib
structs that will need populating soon.