[mtwn] continue implementing the RX reponse bits for MCU commands
* extend struct mtwn_cmd with the response buffer state
* allocate an optional response buffer when a cmd is allocated
* add mtwn_usb_cmd_copyin_response() to copy in a response into the
  given mtwn_cmd and set a response status
* extend MCU_SEND_MSG() to take an optional rx buffer / length to
  eventually put the response info in.

Now, this isn't completely linked together yet and tested because
I'm trying to figure out how to properly and race-free implement
that final copy.

eg in the normal path where no timeouts/errors happen, the following
happens:

* mtwn_usb_cmd_get()
* populate it
* queue it with mtwn_usb_cmd_queue_wait()
* that starts a transfer and calls mtwn_usb_cmd_wait() with a timeout
* mtwn_usb_cmd_wait() msleep's on the command buffer "cmd", but it
  no longer owns the buffer, and it can timeout before completing
* ...
* then in the RX path, mtwn_mcu_mt7610u_mcu_handle_response() will
  handle the response, call mtwn_usb_cmd_copyin_response() to copy in the
  response into the mtwn_cmd, then complete it using mtwn_usb_cmd_complete().
* mtwn_usb_cmd_complete() calls wakeup(), then recycles the mtwn_cmd.
* ...
* the sending thread wakes up from its msleep(), so it knows the TX and
  RX has completed without a timeout.

Now, since mtwn_cmd isn't "owned" by the sender anymore - the only rendezvous
is the msleep/wakeup - the only place we could do the copy to the sender
is in mtwn_usb_cmd_complete() which is called by the RX thread.

However, if there's a timeout / error, mtwn_usb_cmd_complete() could be called
after the sender has given up, and any pointer it currently has placed
into mtwn_cmd is no longer valid.

The /right/ way is to force the sender to "own" a buffer being waited on:

* if we hit a timeout, it's just removed from the pending list
* if it's completed in time, it gets found and woken up, and again
  the sender will remove it from the pending list

That way the sender can copy the data out from the cmd buf and then
return it.

I'll .. do that in a subsequent commit.

(see how if_zyd.c does it, and then go fix my if_otus command handling
code to do the right thing as it currently does NOT.)
valid.